support adding & removing Organisms by Organisms + other minor changes

+++ src/organisms/species.rs
@@ -5,10 +5,10 @@ use ggez::graphics::Color;
 #[derive(Clone)]
 pub struct Species {
     pub name: String,
-    pub max_energy: u32,
-    pub max_health: u32,
+    pub max_energy: f32,
+    pub max_health: f32,
     pub max_age: Duration,
-    pub cost_of_birth: u32,
+    pub cost_of_birth: f32,
     pub walk_speed_s: f32,
     pub can_eat_organisms: bool,
     pub photosynthesis_rate_s: f32,
